Fix listing of unknown drives
The listing of unknown devices was done within the subroutine
checkMissingDevicesDB which was called with the list of
devices from the command line and the list $devices_a
from the already processed subroutine checkWhichDevice.
$devices_a contains only devices which were found in the database.
The Problem is that the names of the found devices
could change within checkWhichDevice (to support RAID controller),
therefore the comparrison of the two lists wasn't possible anymore.

This patch creates another list within checkWhichDevice which contains
those devices missing in the database. If this list is empty there is
simply no missing device. So there is no need for another subroutine
to check for missing devices.