Move attribute renaming to it's own method
As multiple duplicate code lines were present in the plugin, the
regular expressions for prepending an attribute name with it's
device was moved to a separate method. There unwanted chars in attribute
names can also be replaced and the single method can be called to get
the final sanitized attribute name.