Add SMART overall-health self-assessment test result
This is the first step to monitor the self-test results of a device.
The result is parsed and added as ID 1025 to the smart attributes array
of the device. 0 states the test has passed and 1 states the test has
failed. The next step is to add 1025 as an attribute to the smartdb JSON
file (as done with 1024 - ATA error count).